% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use-publish.R
\name{use_publish_workflow}
\alias{use_publish_workflow}
\title{Set up a publication workflow in the current project}
\usage{
use_publish_workflow(justfile = TRUE)
}
\arguments{
\item{justfile}{Copy the \code{justfile} into the project? Set to \code{FALSE} if you
prefer to call \code{pubthis::publish()} directly from R without using \code{just}.}
}
\description{
Copies \code{publish/reference.docx} and \code{publish/docx-format.lua} into the
active project. Optionally also copies a \code{justfile} for terminal-based
publishing.
}
